+++ webapp/assets/_new/styles/tokens.scss
@@ -316,7 +316,7 @@ $z-index-page-submenu: 2500;
 $z-index-page-header: 2000;
 $z-index-page-sidebar: 1500;
 $z-index-sticky: 100;
-$z-index-post-card-link: 5;
+$z-index-post-teaser-link: 5;
 
 /**
  * @tokens Media Query
