+++ webapp/components/PostTeaser/PostTeaser.spec.js
@@ -2,14 +2,14 @@ import { config, shallowMount, mount, RouterLinkStub } from '@vue/test-utils'
 
 import Vuex from 'vuex'
 
-import PostCard from './PostCard.vue'
+import PostTeaser from './PostTeaser.vue'
 
 const localVue = global.localVue
 
 config.stubs['client-only'] = '<span><slot /></span>'
 config.stubs['v-popover'] = '<span><slot /></span>'
 
-describe('PostCard', () => {
+describe('PostTeaser', () => {
   let store
   let stubs
   let mocks
@@ -55,7 +55,7 @@ describe('PostCard', () => {
   describe('shallowMount', () => {
     Wrapper = () => {
       store = new Vuex.Store({ getters })
-      return shallowMount(PostCard, {
+      return shallowMount(PostTeaser, {
         store,
         propsData,
         mocks,
@@ -99,7 +99,7 @@ describe('PostCard', () => {
       const store = new Vuex.Store({
         getters,
       })
-      return mount(PostCard, {
+      return mount(PostTeaser, {
         stubs,
         mocks,
         propsData,
