+++ webapp/pages/index.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <masonry-grid>
-      <ds-grid-item v-show="hashtag" :row-span="2" column-span="fullWidth">
+      <ds-grid-item v-if="hashtag" :row-span="2" column-span="fullWidth">
         <filter-menu :hashtag="hashtag" @clearSearch="clearSearch" />
       </ds-grid-item>
       <ds-grid-item :row-span="2" column-span="fullWidth" class="top-info-bar">
@@ -26,7 +26,7 @@
           :key="post.id"
           :imageAspectRatio="post.imageAspectRatio"
         >
-          <hc-post-card
+          <post-teaser
             :post="post"
             @removePostFromList="deletePost"
             @pinPost="pinPost"
@@ -67,7 +67,7 @@
 // import DonationInfo from '~/components/DonationInfo/DonationInfo.vue'
 import FilterMenu from '~/components/FilterMenu/FilterMenu.vue'
 import HcEmpty from '~/components/Empty/Empty'
-import HcPostCard from '~/components/PostCard/PostCard.vue'
+import PostTeaser from '~/components/PostTeaser/PostTeaser.vue'
 import MasonryGrid from '~/components/MasonryGrid/MasonryGrid.vue'
 import MasonryGridItem from '~/components/MasonryGrid/MasonryGridItem.vue'
 import { mapGetters, mapMutations } from 'vuex'
@@ -79,7 +79,7 @@ export default {
   components: {
     // DonationInfo,
     FilterMenu,
-    HcPostCard,
+    PostTeaser,
     HcEmpty,
     MasonryGrid,
     MasonryGridItem,
